Refactor: move functionality shared by async and sync to shared package
