+++ pydantic_redis/__init__.py
@@ -1,7 +1,9 @@
 """Entry point for redisy"""
 
-from .config import RedisConfig
-from .model import Model
-from .store import Store
+from pydantic_redis.shared.config import RedisConfig
+from pydantic_redis.syncio.model import Model
+from pydantic_redis.syncio.store import Store
+
+__all__ = [Store, RedisConfig, Model]
 
 __version__ = "0.2.0"
