+++ pydantic_redis/shared/model/__init__.py
@@ -1,0 +1,1 @@
+from .base import AbstractModel
