+++ pydantic_redis/syncio/__init__.py
@@ -1,0 +1,1 @@
+"""Package containing the synchronous and thus default version of pydantic_redis"""
